requirements.txt
discord.py
psycopg2-binary
pandas
matplotlib
aiohttp
pillow
